updated tags controller and logic

+++ Services/DB_Services/TagService.cs
@@ -3,7 +3,9 @@ using Microsoft.CodeAnalysis.CSharp.Syntax;
 using Microsoft.EntityFrameworkCore;
 using TaggerApi.DTOs;
 using TaggerApi.Models;
+using TaggerApi.Pagination;
 using TaggerApi.Services.ErrorServices;
+
 
 namespace TaggerApi.Services.DB_Services;
 
@@ -11,19 +13,21 @@ public class TagService : ITagService
 {
 
     private readonly PostgresContext _context;
+    private readonly IPagedList _pageList;
 
-    public TagService(PostgresContext context){
+    public TagService(PostgresContext context, IPagedList pageList){
         _context = context;
+        _pageList = pageList;
     }
 
-    public async Task<TagDTO> AddTag(TagDTO tagDTO)
+    public async Task<TagDTO> AddTag(TagDTO tagDTO,string userUid)
     {
         var  tag = new Tag{
             Id = tagDTO.Id,
             Content = tagDTO.Content,
             Timestamp = tagDTO.Timestamp,
             Medialink = tagDTO.Medialink,
-            IdUser = tagDTO.IdUser,
+            IdUser = userUid,
             IdVideo = tagDTO.IdVideo
         };
 
@@ -33,13 +37,25 @@ public class TagService : ITagService
         return TagToDTO(tag);
     }
 
-    public async Task<bool> DelTag(long id)
+    public async Task<bool> DelTag(long id,string userUid)
     {
         var tag = await _context.Tags.FindAsync(id);
+        
         if (tag == null)
         {
             return false;
         }
+
+        var video = await _context.Videos.FindAsync(tag.IdVideo);
+
+        if (video == null)
+        {
+            throw new NotFoundException("Video not found");
+        }
+         
+        if(userUid != tag.IdUser && userUid != video.IdUser){
+            throw new UnauthorizedAccessException("Tag not your property.");
+        }        
 
         _context.Tags.Remove(tag);
         await _context.SaveChangesAsync();
@@ -47,6 +63,7 @@ public class TagService : ITagService
         return true;        
     }
 
+    //TODO: No use
     public async Task<TagDTO> RetrieveTag(long id)
     {
         var tag = await _context.Tags.FindAsync(id);
@@ -58,7 +75,9 @@ public class TagService : ITagService
 
         return TagToDTO(tag);
     }
+    
 
+    //TODO: No use
     public async Task<IEnumerable<TagDTO>> RetrieveTags()
     {
         return await _context.Tags
@@ -66,29 +85,24 @@ public class TagService : ITagService
             .ToListAsync();
     }
 
-    public async Task<TagDTO> UpdateTag(long id, TagDTO tagDTO)
+
+    public async Task<TagDTO> UpdateTag(long id, TagDTO tagDTO,string userUid)
     {
         var tag = await _context.Tags.FindAsync(id);
 
         if(tag == null){
-            throw new NotFoundException("Video not found.");
+            throw new NotFoundException("Tag not found.");
         }
+
+        if(userUid != tag.IdUser){
+            throw new UnauthorizedAccessException("Tag not your property.");
+        }        
 
         tag.Content = tagDTO.Content;
         tag.Timestamp = tagDTO.Timestamp;
         tag.Medialink = tagDTO.Medialink;
-        tag.IdUser = tagDTO.IdUser;
-        tag.IdVideo = tagDTO.IdVideo;
-
-        try
-        {
-            await _context.SaveChangesAsync();
-        }
-        catch (DbUpdateConcurrencyException)
-        when (!TagExists(id))
-        {
-            throw new NotFoundException("Video not found.");
-        }
+        
+        await _context.SaveChangesAsync();
 
         return TagToDTO(tag);
     }
@@ -105,7 +119,27 @@ public class TagService : ITagService
         Content = tag.Content,
         Timestamp = tag.Timestamp,
         Medialink = tag.Medialink,
-        IdUser = tag.IdUser,
         IdVideo = tag.IdVideo
     };
+
+    public async Task<IEnumerable<TagDTO>> GetTagsPag(PaginationParams request,int idvideo)
+    {
+       var query = _context.Tags.Where(b => b.IdVideo == idvideo);
+
+       var resultPagination = await _pageList.CreatePagedGenericResults<Tag>(query,
+           request.PageNumber,
+           request.PageSize,
+           request.OrderBy!,
+           request.OrderAsc
+       );
+
+       var listDto = new List<TagDTO>();
+       
+       foreach (var item in resultPagination.Result){
+            listDto.Add(TagToDTO(item));
+            Console.WriteLine(item);
+       }              
+
+       return listDto;
+    }
 }
